fix(#40): 필터 스타일 수정
