+++ src/pages/TestPage.tsx
@@ -3,6 +3,7 @@ import useSearchInput from "../hooks/useSearchInput";
 import useModal from "../hooks/useModal";
 import usePopup from "../hooks/usePopup";
 import Comment from "../components/journal/Comment";
+import FilterList from "../components/common/FilterList";
 
 function TestPage() {
   const [, SearchInput] = useSearchInput({
@@ -23,10 +24,10 @@ function TestPage() {
       </Modal>
       <Popup padding="30px 40px">신고가 접수되었습니다.</Popup>
       <SearchInput />
-      여행 일정 페이지
       <button onClick={modalOpen}>모달 open</button>
       <button onClick={popupOpen}>팝업 open</button>
-      <Comment />
+      {/* <Comment /> */}
+      <FilterList filters={["정렬", "지역"]} />
     </PageTemplate>
   );
 }
